backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-socketio==5.10.0
requests==2.31.0
Pillow==10.1.0
PyPDF2==3.0.1
reportlab==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1

